+++ java/src/hype/extended/colorist/HColorTransform.java
@@ -16,8 +16,8 @@ import hype.core.drawable.HDrawable;
 import hype.core.util.HColors;
 
 public class HColorTransform implements HColorist {
-	public float _percA, _percR, _percG, _percB;
-	public int _offsetA, _offsetR, _offsetG, _offsetB;
+	private float _percA, _percR, _percG, _percB;
+	private int _offsetA, _offsetR, _offsetG, _offsetB;
 	private boolean fillFlag, strokeFlag;
 	
 	public HColorTransform() {